refactor: example 패키지 제거, 주석 제거, 암호화 관련 클래스 분리, exception 패키지 생성, shema sql 파일 생성 등
