--- src/main/java/com/example/mybatispractice/MybatisPracticeApplication.java
@@ -1,13 +1,0 @@
-package com.example.mybatispractice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MybatisPracticeApplication {
-
-//  public static void main(String[] args) {
-//    SpringApplication.run(MybatisPracticeApplication.class, args);
-//  }
-
-}
