--- src/main/java/com/example/mybatispractice/domain/model/Person.java
@@ -1,5 +1,0 @@
-package com.example.mybatispractice.domain.model;
-
-public record Person(String id, String name, int age, String address) {
-
-}
