--- src/main/java/com/kjy/trendpop/TrendpopApplication.java
@@ -1,13 +1,0 @@
-package com.kjy.trendpop;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TrendpopApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TrendpopApplication.class, args);
-    }
-
-}
