--- src/main/java/com/kjy/trendpop/config/WebConfig.java
@@ -1,21 +1,0 @@
-package com.kjy.trendpop.config;
-
-import com.kjy.trendpop.sequrity.jwt.JwtInterceptor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Autowired
-    private JwtInterceptor jwtInterceptor;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(jwtInterceptor)
-                .addPathPatterns("/**")  // 모든 경로에 적용
-                .excludePathPatterns("/api/user/authenticate", "/api/user/signup");  // 로그인 및 회원가입 경로는 제외
-    }
-}
