+++ src/main/java/com/trendpop/TrendpopApplication.java
@@ -1,0 +1,13 @@
+package com.trendpop;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TrendpopApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TrendpopApplication.class, args);
+    }
+
+}
