+++ src/main/java/com/trendpop/exception/InvalidCredentialsException.java
@@ -1,0 +1,9 @@
+package com.trendpop.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class InvalidCredentialsException extends BusinessException {
+    public InvalidCredentialsException() {
+        super("INVALID_CREDENTIALS", HttpStatus.UNAUTHORIZED);
+    }
+}
