+++ src/main/java/com/trendpop/exception/UserNotFoundException.java
@@ -1,0 +1,9 @@
+package com.trendpop.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class UserNotFoundException extends BusinessException {
+    public UserNotFoundException() {
+        super("USER_NOT_FOUND", HttpStatus.UNAUTHORIZED);
+    }
+}
