+++ src/main/java/com/trendpop/presentation/dto/request/UserRequest.java
@@ -1,0 +1,26 @@
+package com.trendpop.presentation.dto.request;
+
+import com.trendpop.domain.model.User;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserRequest {
+
+    private String id;
+    private String email;
+    private String password;
+    private String name;
+    private String nickname;
+    private String description;
+    private String profile_image_url;
+    private String phone_number;
+
+    public User toDomain() {
+        return new User(id, email, password, name, nickname, description, profile_image_url, phone_number);
+    }
+
+}
