--- src/main/resources/mapper/PersonMapper.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.mybatispractice.infrastructure.mapper.PersonMapper">
-  <select id="find" parameterType="String" resultType="com.example.mybatispractice.domain.model.Person">
-    SELECT
-      id,
-      name,
-      age,
-      address
-    FROM
-      person
-    WHERE
-      id = #{id}
-  </select>
-
-  <insert id="create" parameterType="com.example.mybatispractice.domain.model.Person" >
-    INSERT INTO
-      person
-    (
-      id,
-      name,
-      age,
-      address
-    )
-    VALUES
-    (
-      #{id},
-      #{name},
-      #{age},
-      #{address}
-    )
-  </insert>
-</mapper>
